Start rework on parser and ir stage.
A program is now defined as a set of functions.
Previously it was only a single expression.
This is to facilitate a change in the x64 part of the compiler regarding blocks and functions.
A function (in the backend) will now contain blocks, instead of a x64 program just being a flat
collection of blocks.